pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voicegateway"
version = "0.0.1"
description = "Self-hosted inference gateway for voice AI — route STT, LLM, and TTS to any provider or local model"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{name = "Mahimai Raja", email = "hello@mahimai.dev"}]
keywords = ["voice-ai", "voicegateway", "inference", "gateway", "self-hosted", "ollama", "livekit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "livekit-agents>=1.5.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
# Cloud providers (install what you need)
deepgram = ["livekit-plugins-deepgram>=1.5.0"]
openai = ["livekit-plugins-openai>=1.5.0"]
anthropic = ["livekit-plugins-anthropic>=1.5.0"]
groq = ["livekit-plugins-openai>=1.5.0"]
cartesia = ["livekit-plugins-cartesia>=1.5.0"]
elevenlabs = ["livekit-plugins-elevenlabs>=1.5.0"]
assemblyai = ["livekit-plugins-assemblyai>=1.5.0"]

# Local models
whisper = ["faster-whisper>=1.0.0"]
kokoro = ["kokoro-onnx>=0.5.0", "onnxruntime>=1.19.0"]
piper = ["piper-tts>=1.2.0"]

# All cloud providers
cloud = ["voicegateway[deepgram,openai,anthropic,groq,cartesia,elevenlabs,assemblyai]"]

# All local models
local = ["voicegateway[whisper,kokoro,piper]"]

# Everything
all = ["voicegateway[cloud,local]"]

# Dashboard
dashboard = ["fastapi>=0.115.0", "uvicorn>=0.32.0"]

# Dev
dev = ["pytest>=8.0", "pytest-asyncio>=0.23.0", "pytest-cov>=5.0"]

[project.scripts]
voicegw = "voicegateway.cli:app"

[project.urls]
Homepage = "https://github.com/mahimailabs/voicegateway"
Documentation = "https://github.com/mahimailabs/voicegateway#readme"
Repository = "https://github.com/mahimailabs/voicegateway"

[tool.setuptools.packages.find]
include = ["voicegateway*", "dashboard*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
